Task_5/requirements.txt
websockets==11.0.3
psycopg2-binary==2.9.7
pandas==2.0.3
asyncio